Refactor insertion points
